requirements.txt
Flask==0.10.1
Flask-MySQL==1.2
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
SQLAlchemy==0.9.8
WTForms==2.0.2
Werkzeug==0.10.1
argparse==1.2.1
arrow==0.5.0
itsdangerous==0.24
pagseguro-python==0.33
python-dateutil==2.4.0
requests==2.5.1
six==1.9.0
voluptuous==0.8.7
wsgiref==0.1.2
wtf==0.1
xmltodict==0.9.2
